+++ app.py
@@ -17506,159 +17506,16 @@ def _get_charts(db: ChDbConnection, dashboard_id: str) -> list[dict]:
     return charts
 
 
-@app.route("/api/dashboards/list", methods=["GET"])
-@require_basic_auth
-async def api_dashboards_list():
-    """Return all non-deleted dashboards for quick picker UIs."""
-    db = get_db()
-    dashboards = _get_dashboards(db)
-    return jsonify({"ok": True, "dashboards": dashboards})
 
 
-@app.route("/api/query/add-to-dashboard", methods=["POST"])
-@require_basic_auth
-async def api_query_add_to_dashboard():
-    """Persist query-page SQL + chart JSON into a dashboard chart record."""
-    payload = await request.get_json(silent=True) or {}
-
-    dashboard_id = str(payload.get("dashboard_id") or "").strip()
-    title = str(payload.get("title") or "").strip()
-    sql = str(payload.get("sql") or "").strip()
-    chart_spec_raw = payload.get("chart_spec")
-
-    if not dashboard_id:
-        return jsonify({"ok": False, "error": "dashboard_id is required"}), 400
-    if not sql:
-        return jsonify({"ok": False, "error": "sql is required"}), 400
-    if not chart_spec_raw:
-        return jsonify({"ok": False, "error": "chart_spec is required"}), 400
-
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        return jsonify({"ok": False, "error": "Dashboard not found"}), 404
-
-    if not title:
-        title = "Query Chart"
-
-    try:
-        chart_option = json.loads(chart_spec_raw) if isinstance(chart_spec_raw, str) else chart_spec_raw
-    except Exception as exc:
-        return jsonify({"ok": False, "error": f"chart_spec must be valid JSON: {exc}"}), 400
-    if not isinstance(chart_option, dict):
-        return jsonify({"ok": False, "error": "chart_spec must be a JSON object"}), 400
-
-    spec_raw = {
-        "template_id": "custom_echarts",
-        "sql": {"mode": "raw", "override_sql": sql},
-        "visual": {
-            "custom_option_json": json.dumps(chart_option, ensure_ascii=False),
-            "custom_mapping_json": "{}",
-        },
-    }
-    try:
-        template_id, query, normalized_spec = _compile_chart_spec(spec_raw)
-    except Exception as exc:
-        return jsonify({"ok": False, "error": f"Chart spec error: {exc}"}), 400
-
-    options_json = json.dumps({"chart_spec": normalized_spec}, ensure_ascii=False)
-    existing = _get_charts(db, dashboard_id)
-    position = max((c["position"] for c in existing), default=-1) + 1
-
-    chart_id = str(uuid.uuid4())
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_chart_configs",
-        [
-            {
-                "Id": chart_id,
-                "DashboardId": dashboard_id,
-                "Title": title,
-                "ChartType": template_id,
-                "Query": query,
-                "OptionsJson": options_json,
-                "Position": position,
-                "IsDeleted": 0,
-                "Version": version,
-            }
-        ],
-    )
-
-    return jsonify(
-        {
-            "ok": True,
-            "chart_id": chart_id,
-            "dashboard_id": dashboard_id,
-            "dashboard_name": dashboard["name"],
-            "dashboard_url": url_for("view_custom_dashboard", dashboard_id=dashboard_id),
-        }
-    )
 
 
-@app.route("/dashboards")
-@require_basic_auth
-async def list_dashboards():
-    db = get_db()
-    dashboards = _get_dashboards(db)
-    return await render_template("custom_dashboards.html", dashboards=dashboards)
 
 
-@app.route("/dashboards/new", methods=["GET"])
-@require_basic_auth
-async def new_dashboard_form():
-    return await render_template("custom_dashboards.html", dashboards=[], show_new_form=True)
 
 
-@app.route("/dashboards", methods=["POST"])
-@require_basic_auth
-async def create_dashboard():
-    form = await request.form
-    name = (form.get("name") or "").strip()
-    description = (form.get("description") or "").strip()
-    if not name:
-        await flash("Dashboard name is required", "warning")
-        return redirect(url_for("list_dashboards"))
-    dashboard_id = str(uuid.uuid4())
-    version = int(time.time() * 1000)
-    db = get_db()
-    _insert_rows_json_each_row(
-        db,
-        "sobs_dashboards",
-        [{"Id": dashboard_id, "Name": name, "Description": description, "IsDeleted": 0, "Version": version}],
-    )
-    return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
 
 
-@app.route("/dashboards/<dashboard_id>")
-@require_basic_auth
-async def view_custom_dashboard(dashboard_id: str):
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        await flash("Dashboard not found", "danger")
-        return redirect(url_for("list_dashboards"))
-    charts = _get_charts(db, dashboard_id)
-    # Convert chart_type to template metadata for frontend
-    templates = [
-        {
-            "id": tid,
-            "name": t["name"],
-            "description": t["description"],
-            "icon": t["icon"],
-            "query_shape": t.get("query_shape", ""),
-            "sample_sql": t.get("sample_sql", ""),
-            "drilldown": t.get("drilldown"),
-            "default_spec": _default_chart_spec(tid),
-        }
-        for tid, t in sorted(CHART_TEMPLATES.items())
-    ]
-    return await render_template(
-        "custom_dashboard_view.html",
-        dashboard=dashboard,
-        charts=charts,
-        templates=templates,
-    )
 
 
 def _register_help_route(path: str, endpoint: str, template_name: str) -> None:
@@ -17733,87 +17590,8 @@ async def _soft_delete_latest_row(
     return redirect(url_for(redirect_endpoint))
 
 
-@app.route("/dashboards/<dashboard_id>/delete", methods=["POST"])
-@require_basic_auth
-async def delete_dashboard(dashboard_id: str):
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        await flash("Dashboard not found", "danger")
-        return redirect(url_for("list_dashboards"))
-    version = int(time.time() * 1000)
-    # Soft-delete dashboard
-    _insert_rows_json_each_row(
-        db,
-        "sobs_dashboards",
-        [
-            {
-                "Id": dashboard_id,
-                "Name": dashboard["name"],
-                "Description": dashboard["description"],
-                "IsDeleted": 1,
-                "Version": version,
-            }
-        ],
-    )
-    # Soft-delete all charts in this dashboard
-    charts = _get_charts(db, dashboard_id)
-    if charts:
-        tombstones = [
-            {
-                "Id": c["id"],
-                "DashboardId": dashboard_id,
-                "Title": c["title"],
-                "ChartType": c["chart_type"],
-                "Query": c["query"],
-                "OptionsJson": c["options_json"],
-                "Position": c["position"],
-                "IsDeleted": 1,
-                "Version": version,
-            }
-            for c in charts
-        ]
-        _insert_rows_json_each_row(db, "sobs_chart_configs", tombstones)
-    await flash(f"Dashboard '{dashboard['name']}' deleted", "success")
-    return redirect(url_for("list_dashboards"))
 
 
-@app.route("/dashboards/<dashboard_id>/charts", methods=["POST"])
-@require_basic_auth
-async def add_chart(dashboard_id: str):
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        await flash("Dashboard not found", "danger")
-        return redirect(url_for("list_dashboards"))
-    form = await request.form
-    try:
-        title, template_id, query, options_json = _parse_chart_form_submission(form)
-    except ValueError as ve:
-        await flash(str(ve), "warning")
-        return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
-    existing = _get_charts(db, dashboard_id)
-    position = max((c["position"] for c in existing), default=-1) + 1
-    chart_id = str(uuid.uuid4())
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_chart_configs",
-        [
-            {
-                "Id": chart_id,
-                "DashboardId": dashboard_id,
-                "Title": title,
-                "ChartType": template_id,
-                "Query": query,
-                "OptionsJson": options_json,
-                "Position": position,
-                "IsDeleted": 0,
-                "Version": version,
-            }
-        ],
-    )
-    return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
 
 
 def _parse_chart_form_submission(form) -> tuple[str, str, str, str]:
@@ -17835,426 +17613,26 @@ def _parse_chart_form_submission(form) -> tuple[str, str, str, str]:
     return title, template_id, query, options_json
 
 
-@app.route("/dashboards/<dashboard_id>/charts/<chart_id>/edit", methods=["POST"])
-@require_basic_auth
-async def edit_chart(dashboard_id: str, chart_id: str):
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        await flash("Dashboard not found", "danger")
-        return redirect(url_for("list_dashboards"))
-
-    charts = _get_charts(db, dashboard_id)
-    chart = next((c for c in charts if c["id"] == chart_id), None)
-    if not chart:
-        await flash("Chart not found", "warning")
-        return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
-
-    form = await request.form
-    try:
-        title, template_id, query, options_json = _parse_chart_form_submission(form)
-    except ValueError as ve:
-        await flash(str(ve), "warning")
-        return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
-
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_chart_configs",
-        [
-            {
-                "Id": chart_id,
-                "DashboardId": dashboard_id,
-                "Title": title,
-                "ChartType": template_id,
-                "Query": query,
-                "OptionsJson": options_json,
-                "Position": chart["position"],
-                "IsDeleted": 0,
-                "Version": version,
-            }
-        ],
-    )
-    return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
 
 
-@app.route("/dashboards/<dashboard_id>/charts/<chart_id>/clone", methods=["POST"])
-@require_basic_auth
-async def clone_chart(dashboard_id: str, chart_id: str):
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        await flash("Dashboard not found", "danger")
-        return redirect(url_for("list_dashboards"))
-
-    charts = _get_charts(db, dashboard_id)
-    source_chart = next((c for c in charts if c["id"] == chart_id), None)
-    if not source_chart:
-        await flash("Chart not found", "warning")
-        return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
-
-    form = await request.form
-    try:
-        title, template_id, query, options_json = _parse_chart_form_submission(form)
-    except ValueError as ve:
-        await flash(str(ve), "warning")
-        return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
-
-    position = max((c["position"] for c in charts), default=-1) + 1
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_chart_configs",
-        [
-            {
-                "Id": str(uuid.uuid4()),
-                "DashboardId": dashboard_id,
-                "Title": title,
-                "ChartType": template_id,
-                "Query": query,
-                "OptionsJson": options_json,
-                "Position": position,
-                "IsDeleted": 0,
-                "Version": version,
-            }
-        ],
-    )
-    return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
 
 
-@app.route("/dashboards/<dashboard_id>/charts/<chart_id>/delete", methods=["POST"])
-@require_basic_auth
-async def remove_chart(dashboard_id: str, chart_id: str):
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        await flash("Dashboard not found", "danger")
-        return redirect(url_for("list_dashboards"))
-    charts = _get_charts(db, dashboard_id)
-    chart = next((c for c in charts if c["id"] == chart_id), None)
-    if not chart:
-        await flash("Chart not found", "warning")
-        return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_chart_configs",
-        [
-            {
-                "Id": chart_id,
-                "DashboardId": dashboard_id,
-                "Title": chart["title"],
-                "ChartType": chart["chart_type"],
-                "Query": chart["query"],
-                "OptionsJson": chart["options_json"],
-                "Position": chart["position"],
-                "IsDeleted": 1,
-                "Version": version,
-            }
-        ],
-    )
-    return redirect(url_for("view_custom_dashboard", dashboard_id=dashboard_id))
 
 
-@app.route("/api/dashboards/query", methods=["POST"])
-@require_basic_auth
-async def execute_chart_query():
-    """Execute a ClickHouse SELECT query and return raw results for eChart rendering."""
-    body = await request.get_json(silent=True) or {}
-    query = (body.get("query") or "").strip()
-    err = _validate_chart_query(query)
-    if err:
-        return jsonify({"error": err}), 400
-    # Inject a row limit to prevent runaway queries
-    if not re.search(r"\bLIMIT\b", query, re.IGNORECASE):
-        query = query.rstrip(";") + " LIMIT 1000"
-    db = get_db()
-    try:
-        result = db.execute(query)
-        rows = result.fetchall()
-        columns = list(rows[0].keys()) if rows else []
-        data = [[row[col] for col in columns] for row in rows]
-        return jsonify({"columns": columns, "rows": data})
-    except Exception as exc:
-        app.logger.exception("Chart query execution failed: %s", query)
-        return jsonify({"error": _public_dashboard_query_error(exc)}), 400
 
 
-@app.route("/api/dashboards/spec/templates", methods=["GET"])
-@require_basic_auth
-async def list_chart_spec_templates():
-    templates = [
-        {
-            "id": tid,
-            "name": t["name"],
-            "description": t["description"],
-            "query_shape": t.get("query_shape", ""),
-            "sample_sql": t.get("sample_sql", ""),
-            "default_spec": _default_chart_spec(tid),
-            "min_columns": t.get("min_columns", 0),
-            "max_columns": t.get("max_columns"),
-            "column_roles": t.get("column_roles", {}),
-        }
-        for tid, t in sorted(CHART_TEMPLATES.items())
-    ]
-    return jsonify({"templates": templates})
 
 
-@app.route("/api/dashboards/spec/options", methods=["GET"])
-@require_basic_auth
-async def chart_spec_options_api():
-    source_view = str(request.args.get("source_view") or "v_derived_signals_anomaly").strip()
-    signal_source = str(request.args.get("signal_source") or "").strip()
-    limit = _coerce_positive_int(request.args.get("limit"), 100, 1, 500)
-
-    supported_sources = {
-        "v_derived_signals_anomaly",
-        "v_otel_metrics_anomaly",
-        "otel_metrics_gauge",
-        "otel_metrics_sum",
-        "otel_metrics_histogram",
-        "otel_logs",
-        "otel_traces",
-        "sobs_error_resolutions",
-    }
-    if source_view not in supported_sources:
-        return jsonify({"error": "Unsupported source for options"}), 400
-
-    db = get_db()
-
-    def _distinct_values(query: str) -> list[str]:
-        rows = db.execute(query).fetchall()
-        values: list[str] = []
-        for row in rows:
-            val = str(row["v"] or "").strip()
-            if val:
-                values.append(val)
-        return values
-
-    services: list[str] = []
-    signals: list[str] = []
-    metrics: list[str] = []
-
-    if source_view == "v_derived_signals_anomaly":
-        services = _distinct_values(
-            "SELECT DISTINCT ServiceName AS v "
-            "FROM v_derived_signals_anomaly "
-            "WHERE time >= now() - INTERVAL 24 HOUR "
-            "ORDER BY v "
-            f"LIMIT {limit}"
-        )
-        signals = _distinct_values(
-            "SELECT DISTINCT SignalName AS v "
-            "FROM v_derived_signals_anomaly "
-            f"{'WHERE' if signal_source else 'WHERE'} time >= now() - INTERVAL 24 HOUR"
-            + (f" AND SignalSource = {_sql_literal(signal_source)}" if signal_source else "")
-            + " ORDER BY v "
-            f"LIMIT {limit}"
-        )
-    elif source_view in {"otel_logs", "otel_traces"}:
-        services = _distinct_values(
-            "SELECT DISTINCT ServiceName AS v " f"FROM {source_view} " "ORDER BY v " f"LIMIT {limit}"
-        )
-        signals = ["log_volume"] if source_view == "otel_logs" else ["trace_volume"]
-    elif source_view == "sobs_error_resolutions":
-        signals = ["resolved_error_volume"]
-    elif source_view in {"v_otel_metrics_anomaly", "otel_metrics_gauge", "otel_metrics_sum", "otel_metrics_histogram"}:
-        services = _distinct_values(
-            "SELECT DISTINCT ServiceName AS v " f"FROM {source_view} " "ORDER BY v " f"LIMIT {limit}"
-        )
-        metrics = _distinct_values(
-            "SELECT DISTINCT MetricName AS v " f"FROM {source_view} " "ORDER BY v " f"LIMIT {limit}"
-        )
-
-    return jsonify(
-        {
-            "source_view": source_view,
-            "services": services,
-            "signals": signals,
-            "metrics": metrics,
-        }
-    )
 
 
-@app.route("/api/dashboards/spec/compile", methods=["POST"])
-@require_basic_auth
-async def compile_chart_spec_api():
-    body = await request.get_json(silent=True) or {}
-    spec = body.get("spec") if isinstance(body, dict) else {}
-    try:
-        template_id, query, normalized_spec = _compile_chart_spec(spec)
-    except ValueError as ve:
-        return jsonify({"error": str(ve)}), 400
-    except Exception as exc:
-        app.logger.exception("Chart spec compile failed")
-        return jsonify({"error": _public_dashboard_query_error(exc)}), 400
-    return _jsonify_with_optional_sql_output_mask({"template_id": template_id, "query": query, "spec": normalized_spec})
 
 
-@app.route("/api/dashboards/spec/dry-run", methods=["POST"])
-@require_basic_auth
-async def dry_run_chart_spec_api():
-    body = await request.get_json(silent=True) or {}
-    spec = body.get("spec") if isinstance(body, dict) else {}
-    try:
-        template_id, query, normalized_spec = _compile_chart_spec(spec)
-    except ValueError as ve:
-        return jsonify({"error": str(ve)}), 400
-
-    run_query = query
-    if not re.search(r"\bLIMIT\b", run_query, re.IGNORECASE):
-        run_query = run_query.rstrip(";") + " LIMIT 20"
-    db = get_db()
-    try:
-        result = db.execute(run_query)
-        rows = result.fetchall()
-        columns = list(rows[0].keys()) if rows else []
-        data = [[row[col] for col in columns] for row in rows]
-        column_types = _infer_column_types(columns, data)
-    except Exception as exc:
-        app.logger.exception("Chart spec dry-run failed")
-        return jsonify({"error": _public_dashboard_query_error(exc)}), 400
-
-    named_query_results = _execute_chart_spec_named_queries(
-        db,
-        normalized_spec.get("named_queries"),
-        default_limit=5,
-        include_records=False,
-    )
-
-    return _jsonify_with_optional_sql_output_mask(
-        {
-            "template_id": template_id,
-            "query": query,
-            "spec": normalized_spec,
-            "columns": columns,
-            "column_types": column_types,
-            "rows": data,
-            "named_query_results": named_query_results,
-        }
-    )
 
 
-@app.route("/api/dashboards/spec/validate", methods=["POST"])
-@require_basic_auth
-async def validate_chart_spec_api():
-    body = await request.get_json(silent=True) or {}
-    spec = body.get("spec") if isinstance(body, dict) else {}
-    try:
-        template_id, query, normalized_spec = _compile_chart_spec(spec)
-    except ValueError as ve:
-        return jsonify({"valid": False, "error": str(ve)}), 400
-
-    db = get_db()
-    try:
-        run_query = query
-        if not re.search(r"\bLIMIT\b", run_query, re.IGNORECASE):
-            run_query = run_query.rstrip(";") + " LIMIT 200"
-        result = db.execute(run_query)
-        raw_rows = result.fetchall()
-        columns = list(raw_rows[0].keys()) if raw_rows else []
-        data = [dict(row) for row in raw_rows]
-        _render_chart_from_template(template_id, columns, data, normalized_spec)
-    except Exception as exc:
-        return jsonify({"valid": False, "error": _public_dashboard_query_error(exc)}), 400
-
-    return _jsonify_with_optional_sql_output_mask(
-        {
-            "valid": True,
-            "template_id": template_id,
-            "query": query,
-            "spec": normalized_spec,
-            "columns": columns,
-            "row_count": len(data),
-        }
-    )
 
 
-@app.route("/api/dashboards/spec/render", methods=["POST"])
-@require_basic_auth
-async def render_chart_spec_api():
-    body = await request.get_json(silent=True) or {}
-    spec = body.get("spec") if isinstance(body, dict) else {}
-    try:
-        template_id, query, normalized_spec = _compile_chart_spec(spec)
-    except ValueError as ve:
-        return jsonify({"error": str(ve)}), 400
-
-    db = get_db()
-    try:
-        run_query = query
-        if not re.search(r"\bLIMIT\b", run_query, re.IGNORECASE):
-            run_query = run_query.rstrip(";") + " LIMIT 1000"
-        result = db.execute(run_query)
-        raw_rows = result.fetchall()
-        columns = list(raw_rows[0].keys()) if raw_rows else []
-        data = [dict(row) for row in raw_rows]
-
-        # Execute named queries and collect datasets
-        named_datasets: dict[str, dict[str, object]] = {}
-        named_query_results = _execute_chart_spec_named_queries(
-            db,
-            normalized_spec.get("named_queries"),
-            default_limit=1000,
-            include_records=True,
-        )
-        for nq in named_query_results:
-            nq_name = str(nq.get("name") or "").strip()
-            if not nq_name:
-                continue
-            if str(nq.get("error") or ""):
-                app.logger.warning("Named query '%s' failed during render: %s", nq_name, nq.get("error"))
-            named_datasets[nq_name] = {
-                "columns": nq.get("columns") or [],
-                "records": nq.get("records") or [],
-                "rows": nq.get("rows") or [],
-            }
-
-        option = _render_chart_from_template(template_id, columns, data, normalized_spec, named_datasets=named_datasets)
-        option = _apply_chart_spec_visual_overrides(template_id, option, normalized_spec)
-    except Exception as exc:
-        app.logger.exception("Chart spec render failed")
-        return jsonify({"error": _public_dashboard_query_error(exc)}), 400
-    return _jsonify_with_optional_sql_output_mask(
-        {"template_id": template_id, "query": query, "spec": normalized_spec, "option": option}
-    )
 
 
-@app.route("/api/dashboards/render", methods=["POST"])
-@require_basic_auth
-async def render_chart():
-    """Execute a query and render with a template to produce eCharts option."""
-    body = await request.get_json(silent=True) or {}
-    query = (body.get("query") or "").strip()
-    template_id = (body.get("template_id") or "time_series_percentiles").strip()
-
-    err = _validate_chart_query(query)
-    if err:
-        return jsonify({"error": err}), 400
-
-    if template_id not in CHART_TEMPLATES:
-        return jsonify({"error": f"Unknown template: {template_id}"}), 400
-
-    # Inject a row limit to prevent runaway queries
-    if not re.search(r"\bLIMIT\b", query, re.IGNORECASE):
-        query = query.rstrip(";") + " LIMIT 1000"
-
-    db = get_db()
-    try:
-        result = db.execute(query)
-        raw_rows = result.fetchall()
-        columns = list(raw_rows[0].keys()) if raw_rows else []
-        data = [dict(row) for row in raw_rows]
-
-        # Render using template
-        option = _render_chart_from_template(template_id, columns, data)
-        return jsonify({"option": option})
-    except ValueError as ve:
-        # Template column mismatch
-        return jsonify({"error": str(ve)}), 400
-    except Exception as exc:
-        app.logger.exception("Chart render failed: template=%s query=%s", template_id, query)
-        return jsonify({"error": _public_dashboard_query_error(exc)}), 400
 
 
 async def _vanna_validate_and_execute_with_repair(
@@ -18399,326 +17777,12 @@ async def _vanna_execute_named_queries(
     return results
 
 
-def _execute_chart_spec_named_queries(
-    db: "ChDbConnection",
-    named_queries: object,
-    *,
-    default_limit: int,
-    include_records: bool,
-) -> list[dict[str, object]]:
-    """Execute spec named queries with uniform output shape for dry-run/render."""
-    results: list[dict[str, object]] = []
-    if not isinstance(named_queries, list):
-        return results
-    for nq in named_queries:
-        if not isinstance(nq, dict):
-            continue
-        nq_name = str(nq.get("name") or "").strip()
-        nq_sql = str(nq.get("sql") or "").strip()
-        if not nq_name or not nq_sql:
-            continue
-        nq_run = nq_sql if re.search(r"\bLIMIT\b", nq_sql, re.IGNORECASE) else f"{nq_sql} LIMIT {default_limit}"
-        try:
-            nq_result = db.execute(nq_run)
-            nq_rows = nq_result.fetchall()
-            nq_columns = list(nq_rows[0].keys()) if nq_rows else []
-            nq_data = [[row[col] for col in nq_columns] for row in nq_rows]
-            item: dict[str, object] = {
-                "name": nq_name,
-                "purpose": str(nq.get("purpose") or ""),
-                "columns": nq_columns,
-                "rows": nq_data,
-                "error": "",
-            }
-            if include_records:
-                item["records"] = [dict(row) for row in nq_rows]
-            results.append(item)
-        except Exception as exc:
-            item = {
-                "name": nq_name,
-                "purpose": str(nq.get("purpose") or ""),
-                "columns": [],
-                "rows": [],
-                "error": _public_dashboard_query_error(exc),
-            }
-            if include_records:
-                item["records"] = []
-            results.append(item)
-    return results
 
 
-@app.route("/api/dashboards/spec/ai-build", methods=["POST"])
-@require_basic_auth
-async def ai_build_chart_spec():
-    """Generate a dashboard chart spec from a natural-language description using AI.
-
-    Accepts JSON ``{question, preferred_chart_type, chart_instruction, thinking_level}``
-    and returns ``{ok, spec, sql, named_queries, columns}``.
-    """
-    payload = await request.get_json(silent=True) or {}
-    question = str(payload.get("question") or "").strip()
-    preferred_chart_type = str(payload.get("preferred_chart_type") or "").strip()
-    chart_instruction = str(payload.get("chart_instruction") or "").strip()
-    thinking_level = _normalize_thinking_level(str(payload.get("thinking_level") or "off"))
-
-    if not question:
-        return jsonify({"ok": False, "error": "question is required"}), 400
-
-    db = get_db()
-    settings = _load_all_ai_settings(db)
-    endpoint_url = settings.get("ai.endpoint_url", "").strip()
-    model = settings.get("ai.model", "").strip()
-    if not endpoint_url or not model:
-        return jsonify({"ok": False, "error": "AI endpoint not configured. Visit Settings → AI Configuration."}), 503
-
-    # Build schema context in a background thread
-    runner = ChdbSqlRunner(db)
-    schema_context = await asyncio.to_thread(runner.get_schema_context)
-
-    # Generate primary SQL
-    sql, sql_err, _sql_stats = await _vanna_generate_sql(
-        question,
-        schema_context,
-        settings,
-        preferred_chart_type=preferred_chart_type,
-        chart_instruction=chart_instruction,
-        thinking_level=thinking_level,
-    )
-    if sql_err:
-        return jsonify({"ok": False, "error": f"SQL generation failed: {sql_err}"}), 503
-
-    # Validate/execute primary SQL and auto-repair if needed.
-    sql, primary_df, primary_error, sql_retry_count, _ = await _vanna_validate_and_execute_with_repair(
-        db=db,
-        question=question,
-        schema_context=schema_context,
-        initial_sql=sql,
-        settings=settings,
-        thinking_level=thinking_level,
-    )
-    if primary_error or primary_df is None:
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": primary_error or "Generated SQL could not be executed.",
-                    "sql": sql,
-                }
-            ),
-            422,
-        )
-
-    # Primary query data for chart generation context.
-    columns: list[str] = []
-    rows: list[list] = []
-    datasets: list[dict[str, Any]] = []
-    columns = list(primary_df.columns)
-    rows = _json_safe_rows(primary_df.values.tolist()) if not primary_df.empty else []
-    datasets.append(
-        {
-            "name": "main",
-            "purpose": "primary dataset",
-            "sql": sql,
-            "columns": columns,
-            "rows": rows,
-        }
-    )
-
-    # Optionally generate named queries for complex multi-dataset charts
-    named_query_results: list[dict[str, Any]] = []
-    if columns:
-        named_queries_raw, _, _ = await _vanna_generate_named_queries(
-            question=question,
-            schema_context=schema_context,
-            base_sql=sql,
-            settings=settings,
-            preferred_chart_type=preferred_chart_type,
-            chart_instruction=chart_instruction,
-            thinking_level=thinking_level,
-        )
-        named_query_results = await _vanna_execute_named_queries(
-            db=db,
-            named_queries=named_queries_raw,
-            question=question,
-            schema_context=schema_context,
-            settings=settings,
-            thinking_level=thinking_level,
-            use_repair=True,
-        )
-        for nq in named_query_results:
-            if not str(nq.get("error") or ""):
-                datasets.append(
-                    {
-                        "name": str(nq.get("name") or ""),
-                        "purpose": str(nq.get("purpose") or ""),
-                        "sql": str(nq.get("sql") or ""),
-                        "columns": nq.get("columns") or [],
-                        "rows": nq.get("rows") or [],
-                    }
-                )
-
-    # Generate eCharts option JSON via LLM
-    chart_spec_json = ""
-    chart_error = ""
-    custom_mapping_json = "{}"
-    if columns:
-        sample = [dict(zip(columns, r)) for r in rows[:20]]
-        chart_spec_json, chart_error, _ = await _vanna_generate_chart_spec(
-            columns,
-            sample,
-            question,
-            settings,
-            preferred_chart_type=preferred_chart_type,
-            chart_instruction=chart_instruction,
-            named_datasets=datasets,
-            thinking_level=thinking_level,
-        )
-        if chart_spec_json:
-            inferred_mapping = _infer_custom_mapping_from_option(chart_spec_json, columns)
-            custom_mapping_json = json.dumps(inferred_mapping, ensure_ascii=False) if inferred_mapping else "{}"
-        else:
-            # Ensure the UI always gets a usable option JSON even when chart generation fails.
-            chart_spec_json = _build_fallback_custom_option_json()
-            custom_mapping_json = json.dumps({"points": {"from": "rows"}}, ensure_ascii=False)
-            chart_error = (
-                f"{chart_error} Using fallback chart option template."
-                if chart_error
-                else "Chart generation failed; using fallback chart option template."
-            )
-
-    named_queries: list[dict[str, str]] = [
-        {
-            "name": str(nq.get("name") or ""),
-            "sql": str(nq.get("sql") or ""),
-            "purpose": str(nq.get("purpose") or ""),
-        }
-        for nq in named_query_results
-        if not str(nq.get("error") or "") and nq.get("name") and nq.get("sql")
-    ]
-
-    spec: dict[str, object] = {
-        "template_id": "custom_echarts",
-        "sql": {"mode": "raw", "override_sql": sql},
-        "named_queries": named_queries,
-        "visual": {
-            "custom_option_json": chart_spec_json or "{}",
-            "custom_mapping_json": custom_mapping_json,
-        },
-    }
-
-    return jsonify(
-        {
-            "ok": True,
-            "spec": spec,
-            "sql": sql,
-            "retry_count": sql_retry_count,
-            "columns": columns,
-            "named_queries": named_queries,
-            "named_query_results": named_query_results,
-            "chart_error": chart_error,
-        }
-    )
 
 
-@app.route("/api/dashboards/<dashboard_id>/charts/<chart_id>/export", methods=["GET"])
-@require_basic_auth
-async def export_chart(dashboard_id: str, chart_id: str):
-    """Export a chart configuration as a downloadable JSON template."""
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        return jsonify({"ok": False, "error": "Dashboard not found"}), 404
-
-    charts = _get_charts(db, dashboard_id)
-    chart = next((c for c in charts if c["id"] == chart_id), None)
-    if not chart:
-        return jsonify({"ok": False, "error": "Chart not found"}), 404
-
-    template_payload: dict[str, object] = {
-        "sobs_chart_template_version": 1,
-        "title": chart["title"],
-        "chart_spec": chart["chart_spec"],
-    }
-
-    safe_title = re.sub(r"[^a-zA-Z0-9_-]", "_", chart["title"])[:64] or "chart"
-    filename = f"sobs_chart_{safe_title}.json"
-    from quart import Response as QuartResponse
-
-    return QuartResponse(
-        json.dumps(template_payload, ensure_ascii=False, indent=2),
-        mimetype="application/json",
-        headers={"Content-Disposition": f'attachment; filename="{filename}"'},
-    )
 
 
-@app.route("/api/dashboards/<dashboard_id>/charts/import", methods=["POST"])
-@require_basic_auth
-async def import_chart(dashboard_id: str):
-    """Import a chart from a JSON template and add it to the dashboard."""
-    db = get_db()
-    dashboard = _get_dashboard(db, dashboard_id)
-    if not dashboard:
-        return jsonify({"ok": False, "error": "Dashboard not found"}), 404
-
-    payload = await request.get_json(silent=True) or {}
-
-    template_version = payload.get("sobs_chart_template_version")
-    if template_version != 1:
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": "Invalid or unsupported chart template format (expected sobs_chart_template_version: 1)",
-                }
-            ),
-            400,
-        )
-
-    title = str(payload.get("title") or "").strip()
-    if not title:
-        title = "Imported Chart"
-
-    chart_spec_raw = payload.get("chart_spec")
-    if not chart_spec_raw:
-        return jsonify({"ok": False, "error": "chart_spec is required in template"}), 400
-
-    try:
-        template_id, query, normalized_spec = _compile_chart_spec(chart_spec_raw)
-    except Exception as exc:
-        return jsonify({"ok": False, "error": f"Chart spec error: {exc}"}), 400
-
-    options_json = json.dumps({"chart_spec": normalized_spec}, ensure_ascii=False)
-    existing = _get_charts(db, dashboard_id)
-    position = max((c["position"] for c in existing), default=-1) + 1
-
-    chart_id_new = str(uuid.uuid4())
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_chart_configs",
-        [
-            {
-                "Id": chart_id_new,
-                "DashboardId": dashboard_id,
-                "Title": title,
-                "ChartType": template_id,
-                "Query": query,
-                "OptionsJson": options_json,
-                "Position": position,
-                "IsDeleted": 0,
-                "Version": version,
-            }
-        ],
-    )
-
-    return jsonify(
-        {
-            "ok": True,
-            "chart_id": chart_id_new,
-            "dashboard_id": dashboard_id,
-            "dashboard_url": url_for("view_custom_dashboard", dashboard_id=dashboard_id),
-        }
-    )
 
 
 # ---------------------------------------------------------------------------
@@ -18822,388 +17886,33 @@ async def metrics_anomaly():
 # ---------------------------------------------------------------------------
 
 # Valid page types for reports
-_REPORT_PAGE_TYPES = {"logs", "traces", "errors", "metrics", "rum", "ai", "work_items", "web_traffic"}
 
 
-def _parse_report_filters(raw_filters_json: Any) -> dict[str, Any]:
-    if not raw_filters_json:
-        return {}
-    try:
-        parsed = json.loads(str(raw_filters_json))
-    except (TypeError, ValueError):
-        return {}
-    return parsed if isinstance(parsed, dict) else {}
 
 
-def _get_reports(db: ChDbConnection, page_type: str | None = None) -> list[dict]:
-    if page_type:
-        rows = db.execute(
-            "SELECT Id, Name, Description, PageType, FiltersJson "
-            "FROM sobs_reports FINAL WHERE IsDeleted = 0 AND PageType = ? ORDER BY Name",
-            [page_type],
-        ).fetchall()
-    else:
-        rows = db.execute(
-            "SELECT Id, Name, Description, PageType, FiltersJson "
-            "FROM sobs_reports FINAL WHERE IsDeleted = 0 ORDER BY PageType, Name"
-        ).fetchall()
-    return [
-        {
-            "id": str(r["Id"]),
-            "name": str(r["Name"]),
-            "description": str(r["Description"]),
-            "page_type": str(r["PageType"]),
-            "filters": _parse_report_filters(r["FiltersJson"]),
-        }
-        for r in rows
-    ]
 
 
-def _get_report(db: ChDbConnection, report_id: str) -> dict | None:
-    row = db.execute(
-        "SELECT Id, Name, Description, PageType, FiltersJson " "FROM sobs_reports FINAL WHERE IsDeleted = 0 AND Id = ?",
-        [report_id],
-    ).fetchone()
-    if not row:
-        return None
-    return {
-        "id": str(row["Id"]),
-        "name": str(row["Name"]),
-        "description": str(row["Description"]),
-        "page_type": str(row["PageType"]),
-        "filters": _parse_report_filters(row["FiltersJson"]),
-    }
 
 
-@app.route("/reports")
-@require_basic_auth
-async def list_reports():
-    db = get_db()
-    reports = _get_reports(db)
-    return await render_template("reports.html", reports=reports)
 
 
-@app.route("/reports/<report_id>/delete", methods=["POST"])
-@require_basic_auth
-async def delete_report(report_id: str):
-    db = get_db()
-    report = _get_report(db, report_id)
-    if not report:
-        await flash("Report not found", "danger")
-        return redirect(url_for("list_reports"))
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_reports",
-        [
-            {
-                "Id": report_id,
-                "Name": report["name"],
-                "Description": report["description"],
-                "PageType": report["page_type"],
-                "FiltersJson": json.dumps(report["filters"], ensure_ascii=False),
-                "IsDeleted": 1,
-                "Version": version,
-            }
-        ],
-    )
-    await flash(f"Report '{report['name']}' deleted", "success")
-    return redirect(url_for("list_reports"))
 
 
-@app.route("/api/reports", methods=["GET"])
-@require_basic_auth
-async def api_list_reports():
-    page_type = request.args.get("page_type", "").strip()
-    db = get_db()
-    reports = _get_reports(db, page_type if page_type else None)
-    return jsonify(reports)
 
 
-@app.route("/api/reports", methods=["POST"])
-@require_basic_auth
-async def api_create_report():
-    body = await request.get_json(silent=True) or {}
-    name = (body.get("name") or "").strip()
-    description = (body.get("description") or "").strip()
-    page_type = (body.get("page_type") or "").strip()
-    filters = body.get("filters") or {}
-
-    if not name:
-        return jsonify({"error": "name is required"}), 400
-    if page_type not in _REPORT_PAGE_TYPES:
-        return jsonify({"error": f"page_type must be one of: {', '.join(sorted(_REPORT_PAGE_TYPES))}"}), 400
-    if not isinstance(filters, dict):
-        return jsonify({"error": "filters must be an object"}), 400
-
-    report_id = str(uuid.uuid4())
-    version = int(time.time() * 1000)
-    db = get_db()
-    _insert_rows_json_each_row(
-        db,
-        "sobs_reports",
-        [
-            {
-                "Id": report_id,
-                "Name": name,
-                "Description": description,
-                "PageType": page_type,
-                "FiltersJson": json.dumps(filters, ensure_ascii=False),
-                "IsDeleted": 0,
-                "Version": version,
-            }
-        ],
-    )
-    result = {"id": report_id, "name": name, "description": description, "page_type": page_type, "filters": filters}
-    return jsonify(result), 201
 
 
-@app.route("/api/reports/<report_id>", methods=["DELETE"])
-@require_basic_auth
-async def api_delete_report(report_id: str):
-    db = get_db()
-    report = _get_report(db, report_id)
-    if not report:
-        return jsonify({"error": "not found"}), 404
-    version = int(time.time() * 1000)
-    _insert_rows_json_each_row(
-        db,
-        "sobs_reports",
-        [
-            {
-                "Id": report_id,
-                "Name": report["name"],
-                "Description": report["description"],
-                "PageType": report["page_type"],
-                "FiltersJson": json.dumps(report["filters"], ensure_ascii=False),
-                "IsDeleted": 1,
-                "Version": version,
-            }
-        ],
-    )
-    return jsonify({"deleted": True})
 
 
 # Export schema version for forward-compatibility
-_REPORTS_EXPORT_VERSION = "1"
 
 # Maximum number of reports that may be imported in a single request
-_REPORTS_IMPORT_MAX = 500
 
 # Maximum raw request body size accepted by report import
-_REPORTS_IMPORT_MAX_BYTES = 5 * 1024 * 1024
 
 
-@app.route("/api/reports/export", methods=["GET"])
-@require_basic_auth
-async def api_export_reports():
-    """Export one or more saved reports as a portable JSON payload.
-
-    Query parameters:
-      ids – comma-separated list of report UUIDs.  Omit to export all reports.
-
-    Returns a ``Content-Disposition: attachment`` JSON file so the browser
-    triggers a download automatically.
-    """
-    db = get_db()
-    raw_ids = request.args.get("ids", "").strip()
-    if raw_ids:
-        wanted = {s.strip() for s in raw_ids.split(",") if s.strip()}
-        all_reports = _get_reports(db)
-        reports = [r for r in all_reports if r["id"] in wanted]
-    else:
-        reports = _get_reports(db)
-
-    payload = {
-        "sobs_reports_export": True,
-        "version": _REPORTS_EXPORT_VERSION,
-        "exported_at": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "reports": [
-            {
-                "id": r["id"],
-                "name": r["name"],
-                "description": r["description"],
-                "page_type": r["page_type"],
-                "filters": r["filters"],
-            }
-            for r in reports
-        ],
-    }
-    json_bytes = json.dumps(payload, indent=2, ensure_ascii=False).encode("utf-8")
-    filename = "sobs_reports_export.json"
-    return Response(
-        json_bytes,
-        status=200,
-        headers={
-            "Content-Type": "application/json; charset=utf-8",
-            "Content-Disposition": f'attachment; filename="{filename}"',
-        },
-    )
 
 
-@app.route("/api/reports/import", methods=["POST"])
-@require_basic_auth
-async def api_import_reports():
-    """Import reports from a previously exported JSON payload.
-
-    Accepts either ``application/json`` body or a ``multipart/form-data``
-    upload with a ``file`` field containing the JSON file.
-
-    Body / file must be the JSON object produced by ``/api/reports/export``.
-
-    Optional query / body parameter ``on_conflict`` controls duplicate
-    handling when a report with the same name already exists for the same
-    page_type:
-      - ``rename``  (default) – append " (imported)" / " (2)" etc. to the name.
-      - ``replace`` – delete the existing report and insert the new one.
-      - ``skip``    – leave the existing report unchanged and skip the new one.
-
-    Returns a JSON summary: ``{imported, skipped, replaced, errors}``.
-    """
-
-    def _payload_too_large_response():
-        return jsonify({"error": f"Import payload too large (max {_REPORTS_IMPORT_MAX_BYTES} bytes)"}), 413
-
-    if (request.content_length or 0) > _REPORTS_IMPORT_MAX_BYTES:
-        return _payload_too_large_response()
-
-    # ── Parse body ────────────────────────────────────────────────────────────
-    content_type = request.content_type or ""
-    on_conflict = (request.args.get("on_conflict") or "").strip().lower()
-
-    if "multipart/form-data" in content_type or "application/x-www-form-urlencoded" in content_type:
-        files = await request.files
-        forms = await request.form
-        if not on_conflict:
-            on_conflict = forms.get("on_conflict", "rename").strip().lower()
-        uploaded = files.get("file")
-        if not uploaded:
-            return jsonify({"error": "No file uploaded"}), 400
-        try:
-            raw_bytes = uploaded.read()
-            if len(raw_bytes) > _REPORTS_IMPORT_MAX_BYTES:
-                return _payload_too_large_response()
-            body = json.loads(raw_bytes)
-        except (ValueError, TypeError):
-            return jsonify({"error": "Invalid JSON file"}), 400
-    else:
-        body = await request.get_json(silent=True)
-        if body is None:
-            return jsonify({"error": "Invalid or missing JSON body"}), 400
-        if not on_conflict:
-            on_conflict = (body.get("on_conflict") or "rename").strip().lower()
-
-    # ── Validate envelope ────────────────────────────────────────────────────
-    if not isinstance(body, dict) or not body.get("sobs_reports_export"):
-        return jsonify({"error": "Not a valid SOBS reports export file"}), 400
-    if str(body.get("version", "")) != _REPORTS_EXPORT_VERSION:
-        return jsonify({"error": f"Unsupported export version: {body.get('version')!r}"}), 400
-    if on_conflict not in ("rename", "replace", "skip"):
-        return jsonify({"error": "on_conflict must be one of: rename, replace, skip"}), 400
-
-    incoming = body.get("reports")
-    if not isinstance(incoming, list):
-        return jsonify({"error": "'reports' must be a list"}), 400
-    if len(incoming) > _REPORTS_IMPORT_MAX:
-        return jsonify({"error": f"Too many reports (max {_REPORTS_IMPORT_MAX})"}), 400
-
-    # ── Build index of existing reports by (page_type, lower(name)) ──────────
-    db = get_db()
-    existing = _get_reports(db)
-    existing_index: dict[tuple[str, str], dict] = {(r["page_type"], r["name"].lower()): r for r in existing}
-
-    n_imported = 0
-    n_skipped = 0
-    n_replaced = 0
-    n_errors = 0
-    version_base = int(time.time() * 1000)
-
-    for idx, item in enumerate(incoming):
-        if not isinstance(item, dict):
-            n_errors += 1
-            continue
-        name = (item.get("name") or "").strip()
-        description = (item.get("description") or "").strip()
-        page_type = (item.get("page_type") or "").strip()
-        filters = item.get("filters") or {}
-
-        if not name:
-            n_errors += 1
-            continue
-        if page_type not in _REPORT_PAGE_TYPES:
-            n_errors += 1
-            continue
-        if not isinstance(filters, dict):
-            n_errors += 1
-            continue
-
-        conflict_key = (page_type, name.lower())
-        conflict = existing_index.get(conflict_key)
-
-        if conflict:
-            if on_conflict == "skip":
-                n_skipped += 1
-                continue
-            elif on_conflict == "replace":
-                # Soft-delete the existing report
-                _insert_rows_json_each_row(
-                    db,
-                    "sobs_reports",
-                    [
-                        {
-                            "Id": conflict["id"],
-                            "Name": conflict["name"],
-                            "Description": conflict["description"],
-                            "PageType": conflict["page_type"],
-                            "FiltersJson": json.dumps(conflict["filters"], ensure_ascii=False),
-                            "IsDeleted": 1,
-                            "Version": version_base + idx * 2,
-                        }
-                    ],
-                )
-                n_replaced += 1
-                # Remove from index so we don't double-delete
-                del existing_index[conflict_key]
-            else:
-                # rename – find a unique name
-                candidate = f"{name} (imported)"
-                suffix = 2
-                while (page_type, candidate.lower()) in existing_index:
-                    candidate = f"{name} (imported {suffix})"
-                    suffix += 1
-                name = candidate
-
-        new_id = str(uuid.uuid4())
-        _insert_rows_json_each_row(
-            db,
-            "sobs_reports",
-            [
-                {
-                    "Id": new_id,
-                    "Name": name,
-                    "Description": description,
-                    "PageType": page_type,
-                    "FiltersJson": json.dumps(filters, ensure_ascii=False),
-                    "IsDeleted": 0,
-                    "Version": version_base + idx * 2 + 1,
-                }
-            ],
-        )
-        # Track freshly-imported name to avoid collisions within the same batch
-        existing_index[(page_type, name.lower())] = {"id": new_id, "name": name, "page_type": page_type}
-        if conflict and on_conflict == "replace":
-            # Replacement inserts are counted as replaced, not imported.
-            continue
-        n_imported += 1
-
-    return jsonify(
-        {
-            "imported": n_imported,
-            "skipped": n_skipped,
-            "replaced": n_replaced,
-            "errors": n_errors,
-        }
-    )
 
 
 # ---------------------------------------------------------------------------
@@ -23806,699 +22515,14 @@ def _vanna_run_query(db: "ChDbConnection", sql: str) -> tuple["pd.DataFrame | No
 # ---------------------------------------------------------------------------
 
 
-@app.route("/query")
-@require_basic_auth
-async def view_query():
-    if not _query_page_enabled():
-        return (
-            "Query page is unavailable until AI and guard settings are configured.",
-            404,
-        )
-    return await render_template("query.html")
 
 
-@app.route("/api/query/ask", methods=["POST"])
-@require_basic_auth
-async def api_query_ask():
-    """Natural-language → SQL → DataFrame endpoint.
-
-    Accepts JSON ``{question, execute, chart}`` and returns::
-
-        {
-          ok: bool,
-                    trace_id: str,
-                    turn_id: str,
-          sql: str,
-          columns: [...],
-                    field_types: [{name, dtype, kind}, ...],
-          rows: [[...], ...],
-                    retry_count: int,
-          chart_spec: str,   # ECharts option JSON, may be empty
-          error: str
-        }
-    """
-    payload = await request.get_json(force=True, silent=True) or {}
-    question = str(payload.get("question") or "").strip()
-    do_execute = bool(payload.get("execute", True))
-    do_chart = bool(payload.get("chart", False))
-    preferred_chart_type = str(payload.get("preferred_chart_type") or "").strip()
-    chart_instruction = str(payload.get("chart_instruction") or "").strip()
-    thinking_level = _normalize_thinking_level(str(payload.get("thinking_level") or "off"))
-
-    if not question:
-        return jsonify({"ok": False, "error": "question is required"}), 400
-
-    db = get_db()
-    settings = _load_all_ai_settings(db)
-    if not _query_page_enabled(settings):
-        return jsonify({"ok": False, "error": "Query page is unavailable."}), 404
-
-    trace_id = hashlib.md5(f"query|{question}|{time.time_ns()}".encode("utf-8")).hexdigest()
-    turn_id = trace_id[:16]
-    model = settings.get("ai.model", "").strip()
-    guard_model = settings.get("ai.guard_model", "").strip()
-
-    _emit_ai_helper_log_event(
-        event_name="query.turn.start",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body=question,
-        attrs={"gen_ai.input.question": question},
-    )
-
-    allowed, guard_reason, guard_stats = await _check_guard_model(settings, question, "/query")
-    _emit_ai_helper_log_event(
-        event_name="query.guard.result",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body=f"Guard verdict: {guard_reason}",
-        attrs=_guard_telemetry_attrs(allowed, guard_reason, guard_stats),
-    )
-    if not allowed:
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": f"Request blocked by safety guard: {guard_reason}",
-                    "trace_id": trace_id,
-                    "turn_id": turn_id,
-                }
-            ),
-            403,
-        )
-
-    # Build schema context (run synchronously in a thread so we don't block the event loop)
-    runner = ChdbSqlRunner(db)
-    schema_context = await asyncio.to_thread(runner.get_schema_context)
-
-    # Generate SQL
-    sql, sql_err, sql_stats = await _vanna_generate_sql(
-        question,
-        schema_context,
-        settings,
-        preferred_chart_type=preferred_chart_type,
-        chart_instruction=chart_instruction,
-        thinking_level=thinking_level,
-    )
-    _emit_ai_helper_log_event(
-        event_name="query.sql.generated",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body=sql if sql else sql_err,
-        attrs={
-            "gen_ai.operation.name": "query_sql",
-            "gen_ai.usage.input_tokens": sql_stats.get("prompt_tokens", 0),
-            "gen_ai.usage.output_tokens": sql_stats.get("completion_tokens", 0),
-            "gen_ai.response.latency_ms": sql_stats.get("elapsed_ms", 0),
-            "sobs.gen_ai.prompt": question,
-            "sobs.gen_ai.response": sql,
-        },
-    )
-    if sql_err:
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": sql_err,
-                    "trace_id": trace_id,
-                    "turn_id": turn_id,
-                    "sql": "",
-                    "columns": [],
-                    "rows": [],
-                    "llm_stats": _summarize_query_llm_stats(sql_generation=sql_stats),
-                }
-            ),
-            503,
-        )
-
-    # Optionally execute
-    columns: list[str] = []
-    field_types: list[dict[str, str]] = []
-    rows: list[list] = []
-    datasets: list[dict[str, Any]] = []
-    retry_count = 0
-    exec_error = ""
-    last_repair_stats: dict[str, Any] = {}
-    named_stats: dict[str, Any] = {}
-    chart_stats: dict[str, Any] = {}
-    if do_execute:
-        exec_started = time.monotonic()
-        sql, main_df, exec_error, retry_count, last_repair_stats = await _vanna_validate_and_execute_with_repair(
-            db=db,
-            question=question,
-            schema_context=schema_context,
-            initial_sql=sql,
-            settings=settings,
-            thinking_level=thinking_level,
-        )
-        exec_elapsed_ms = int((time.monotonic() - exec_started) * 1000)
-        row_count = int(len(main_df)) if main_df is not None else 0
-        _emit_ai_helper_log_event(
-            event_name="query.sql.executed",
-            chat_id=trace_id,
-            turn_id=turn_id,
-            page="/query",
-            model=model,
-            guard_model=guard_model,
-            thinking_level="off",
-            body=sql,
-            severity="INFO" if not exec_error else "ERROR",
-            attrs={
-                "gen_ai.operation.name": "query_sql_execute",
-                "sobs.query.exec.attempt": max(1, retry_count + 1),
-                "sobs.query.exec.status": "ok" if not exec_error else "error",
-                "sobs.query.exec.row_count": row_count,
-                "sobs.query.exec.error": exec_error,
-                "gen_ai.response.latency_ms": exec_elapsed_ms,
-                "sobs.gen_ai.prompt": question,
-                "sobs.gen_ai.response": sql,
-            },
-        )
-
-        if main_df is not None and not exec_error:
-            if not main_df.empty:
-                columns = list(main_df.columns)
-                field_types = _infer_query_field_types(main_df)
-                rows = _json_safe_rows(main_df.values.tolist())
-            datasets.append(
-                {
-                    "name": "main",
-                    "purpose": "primary dataset",
-                    "sql": sql,
-                    "columns": columns,
-                    "field_types": field_types,
-                    "rows": rows,
-                    "error": "",
-                }
-            )
-
-    # Optionally generate chart spec
-    chart_spec = ""
-    chart_error = ""
-    if do_chart and not exec_error and columns:
-        named_queries, _named_err, named_stats = await _vanna_generate_named_queries(
-            question=question,
-            schema_context=schema_context,
-            base_sql=sql,
-            settings=settings,
-            preferred_chart_type=preferred_chart_type,
-            chart_instruction=chart_instruction,
-            thinking_level=thinking_level,
-        )
-        _emit_ai_helper_log_event(
-            event_name="query.sql.named_generated",
-            chat_id=trace_id,
-            turn_id=turn_id,
-            page="/query",
-            model=model,
-            guard_model=guard_model,
-            thinking_level="off",
-            body=json.dumps(named_queries, ensure_ascii=False),
-            attrs={
-                "gen_ai.operation.name": "query_sql_named",
-                "gen_ai.usage.input_tokens": named_stats.get("prompt_tokens", 0),
-                "gen_ai.usage.output_tokens": named_stats.get("completion_tokens", 0),
-                "gen_ai.response.latency_ms": named_stats.get("elapsed_ms", 0),
-            },
-        )
-
-        named_results = await _vanna_execute_named_queries(
-            db=db,
-            named_queries=named_queries,
-            question=question,
-            schema_context=schema_context,
-            settings=settings,
-            thinking_level=thinking_level,
-            include_field_types=True,
-            use_repair=False,
-        )
-        for ds in named_results:
-            datasets.append(
-                {
-                    "name": str(ds.get("name") or "dataset"),
-                    "purpose": str(ds.get("purpose") or ""),
-                    "sql": str(ds.get("sql") or ""),
-                    "columns": ds.get("columns") or [],
-                    "field_types": ds.get("field_types") or [],
-                    "rows": ds.get("rows") or [],
-                    "error": str(ds.get("error") or ""),
-                }
-            )
-
-        sample = [dict(zip(columns, r)) for r in rows[:20]]
-        chart_spec, chart_error, chart_stats = await _vanna_generate_chart_spec(
-            columns,
-            sample,
-            question,
-            settings,
-            preferred_chart_type=preferred_chart_type,
-            chart_instruction=chart_instruction,
-            named_datasets=datasets,
-            thinking_level=thinking_level,
-        )
-        _emit_ai_helper_log_event(
-            event_name="query.chart.generated",
-            chat_id=trace_id,
-            turn_id=turn_id,
-            page="/query",
-            model=model,
-            guard_model=guard_model,
-            thinking_level="off",
-            body=chart_spec if chart_spec else chart_error,
-            attrs={
-                "gen_ai.operation.name": "query_chart",
-                "gen_ai.usage.input_tokens": chart_stats.get("prompt_tokens", 0),
-                "gen_ai.usage.output_tokens": chart_stats.get("completion_tokens", 0),
-                "gen_ai.response.latency_ms": chart_stats.get("elapsed_ms", 0),
-            },
-        )
-
-    _emit_ai_helper_log_event(
-        event_name="query.turn.complete",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body="Query turn completed",
-        attrs={
-            "gen_ai.input.question": question,
-            "sobs.gen_ai.prompt": question,
-            "sobs.gen_ai.response": sql,
-            "gen_ai.operation.name": "query",
-        },
-    )
-
-    return _jsonify_with_optional_sql_output_mask(
-        {
-            "ok": True,
-            "trace_id": trace_id,
-            "turn_id": turn_id,
-            "sql": sql,
-            "columns": columns,
-            "field_types": field_types,
-            "rows": rows,
-            "retry_count": retry_count,
-            "datasets": datasets,
-            "chart_spec": chart_spec,
-            "error": exec_error or chart_error,
-            "llm_stats": _summarize_query_llm_stats(
-                sql_generation=sql_stats,
-                sql_repair=last_repair_stats,
-                named_query_generation=named_stats,
-                chart_generation=chart_stats,
-            ),
-        }
-    )
 
 
-@app.route("/api/query/run", methods=["POST"])
-@require_basic_auth
-async def api_query_run():
-    """Execute an existing SQL statement and optionally generate a chart."""
-    payload = await request.get_json(force=True, silent=True) or {}
-    sql = str(payload.get("sql") or "").strip()
-    question = str(payload.get("question") or "").strip()
-    do_chart = bool(payload.get("chart", False))
-    preferred_chart_type = str(payload.get("preferred_chart_type") or "").strip()
-    chart_instruction = str(payload.get("chart_instruction") or "").strip()
-    thinking_level = _normalize_thinking_level(str(payload.get("thinking_level") or "off"))
-
-    if not sql:
-        return jsonify({"ok": False, "error": "sql is required"}), 400
-
-    db = get_db()
-    settings = _load_all_ai_settings(db)
-    if not _query_page_enabled(settings):
-        return jsonify({"ok": False, "error": "Query page is unavailable."}), 404
-
-    trace_id = hashlib.md5(f"query-run|{sql}|{time.time_ns()}".encode("utf-8")).hexdigest()
-    turn_id = trace_id[:16]
-    model = settings.get("ai.model", "").strip()
-    guard_model = settings.get("ai.guard_model", "").strip()
-
-    _emit_ai_helper_log_event(
-        event_name="query.turn.start",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body=question or sql,
-        attrs={"gen_ai.input.question": question or "(manual SQL execution)"},
-    )
-
-    exec_started = time.monotonic()
-    # Pre-flight EXPLAIN to surface any parse/planning errors before execution.
-    explain_error = await asyncio.to_thread(_vanna_explain_sql, db, sql)
-    if explain_error:
-        exec_elapsed_ms = int((time.monotonic() - exec_started) * 1000)
-        _emit_ai_helper_log_event(
-            event_name="query.sql.explain_failed",
-            chat_id=trace_id,
-            turn_id=turn_id,
-            page="/query",
-            model=model,
-            guard_model=guard_model,
-            thinking_level="off",
-            body=explain_error,
-            severity="WARN",
-            attrs={"gen_ai.operation.name": "query_sql_explain", "sobs.query.exec.error": explain_error},
-        )
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": explain_error,
-                    "trace_id": trace_id,
-                    "turn_id": turn_id,
-                    "sql": sql,
-                    "columns": [],
-                    "rows": [],
-                    "llm_stats": _summarize_query_llm_stats(),
-                }
-            ),
-            422,
-        )
-    try:
-        df, exec_error = await asyncio.to_thread(_vanna_run_query, db, sql)
-    except Exception as exc:
-        df, exec_error = None, f"Query execution error: {exc}"
-    exec_elapsed_ms = int((time.monotonic() - exec_started) * 1000)
-
-    row_count = 0
-    columns: list[str] = []
-    field_types: list[dict[str, str]] = []
-    rows: list[list] = []
-    datasets: list[dict[str, Any]] = []
-    if df is not None:
-        row_count = int(len(df))
-        if not df.empty:
-            columns = list(df.columns)
-            field_types = _infer_query_field_types(df)
-            rows = _json_safe_rows(df.values.tolist())
-        datasets.append(
-            {
-                "name": "main",
-                "purpose": "primary dataset",
-                "sql": sql,
-                "columns": columns,
-                "field_types": field_types,
-                "rows": rows,
-                "error": "",
-            }
-        )
-
-    _emit_ai_helper_log_event(
-        event_name="query.sql.executed",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body=sql,
-        severity="INFO" if not exec_error else "ERROR",
-        attrs={
-            "gen_ai.operation.name": "query_sql_execute",
-            "sobs.query.exec.attempt": 1,
-            "sobs.query.exec.status": "ok" if not exec_error else "error",
-            "sobs.query.exec.row_count": row_count,
-            "sobs.query.exec.error": exec_error,
-            "gen_ai.response.latency_ms": exec_elapsed_ms,
-            "sobs.gen_ai.prompt": question,
-            "sobs.gen_ai.response": sql,
-        },
-    )
-
-    chart_spec = ""
-    chart_error = ""
-    named_stats: dict[str, Any] = {}
-    chart_stats: dict[str, Any] = {}
-    if do_chart and not exec_error and columns:
-        guard_input = question or f"Generate chart for SQL: {sql[:500]}"
-        allowed, guard_reason, guard_stats = await _check_guard_model(settings, guard_input, "/query")
-        _emit_ai_helper_log_event(
-            event_name="query.guard.result",
-            chat_id=trace_id,
-            turn_id=turn_id,
-            page="/query",
-            model=model,
-            guard_model=guard_model,
-            thinking_level="off",
-            body=f"Guard verdict: {guard_reason}",
-            attrs=_guard_telemetry_attrs(allowed, guard_reason, guard_stats),
-        )
-        if allowed:
-            schema_context = await asyncio.to_thread(ChdbSqlRunner(db).get_schema_context)
-            named_queries, _named_err, named_stats = await _vanna_generate_named_queries(
-                question=question or sql,
-                schema_context=schema_context,
-                base_sql=sql,
-                settings=settings,
-                preferred_chart_type=preferred_chart_type,
-                chart_instruction=chart_instruction,
-                thinking_level=thinking_level,
-            )
-            _emit_ai_helper_log_event(
-                event_name="query.sql.named_generated",
-                chat_id=trace_id,
-                turn_id=turn_id,
-                page="/query",
-                model=model,
-                guard_model=guard_model,
-                thinking_level="off",
-                body=json.dumps(named_queries, ensure_ascii=False),
-                attrs={
-                    "gen_ai.operation.name": "query_sql_named",
-                    "gen_ai.usage.input_tokens": named_stats.get("prompt_tokens", 0),
-                    "gen_ai.usage.output_tokens": named_stats.get("completion_tokens", 0),
-                    "gen_ai.response.latency_ms": named_stats.get("elapsed_ms", 0),
-                },
-            )
-
-            named_results = await _vanna_execute_named_queries(
-                db=db,
-                named_queries=named_queries,
-                question=question or sql,
-                schema_context=schema_context,
-                settings=settings,
-                thinking_level=thinking_level,
-                include_field_types=True,
-                use_repair=False,
-            )
-            for ds in named_results:
-                datasets.append(
-                    {
-                        "name": str(ds.get("name") or "dataset"),
-                        "purpose": str(ds.get("purpose") or ""),
-                        "sql": str(ds.get("sql") or ""),
-                        "columns": ds.get("columns") or [],
-                        "field_types": ds.get("field_types") or [],
-                        "rows": ds.get("rows") or [],
-                        "error": str(ds.get("error") or ""),
-                    }
-                )
-
-            sample = [dict(zip(columns, r)) for r in rows[:20]]
-            chart_spec, chart_error, chart_stats = await _vanna_generate_chart_spec(
-                columns,
-                sample,
-                question,
-                settings,
-                preferred_chart_type=preferred_chart_type,
-                chart_instruction=chart_instruction,
-                named_datasets=datasets,
-                thinking_level=thinking_level,
-            )
-            _emit_ai_helper_log_event(
-                event_name="query.chart.generated",
-                chat_id=trace_id,
-                turn_id=turn_id,
-                page="/query",
-                model=model,
-                guard_model=guard_model,
-                thinking_level="off",
-                body=chart_spec if chart_spec else chart_error,
-                attrs={
-                    "gen_ai.operation.name": "query_chart",
-                    "gen_ai.usage.input_tokens": chart_stats.get("prompt_tokens", 0),
-                    "gen_ai.usage.output_tokens": chart_stats.get("completion_tokens", 0),
-                    "gen_ai.response.latency_ms": chart_stats.get("elapsed_ms", 0),
-                },
-            )
-        else:
-            chart_error = f"Chart generation blocked by safety guard: {guard_reason}"
-
-    final_error = exec_error or chart_error
-    _emit_ai_helper_log_event(
-        event_name="query.turn.complete",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model=guard_model,
-        thinking_level="off",
-        body="Query turn completed",
-        severity="INFO" if not final_error else "ERROR",
-        attrs={
-            "gen_ai.input.question": question,
-            "sobs.gen_ai.prompt": question,
-            "sobs.gen_ai.response": sql,
-            "gen_ai.operation.name": "query",
-        },
-    )
-
-    return _jsonify_with_optional_sql_output_mask(
-        {
-            "ok": True,
-            "trace_id": trace_id,
-            "turn_id": turn_id,
-            "sql": sql,
-            "columns": columns,
-            "field_types": field_types,
-            "rows": rows,
-            "retry_count": 0,
-            "datasets": datasets,
-            "chart_spec": chart_spec,
-            "error": final_error,
-            "llm_stats": _summarize_query_llm_stats(
-                named_query_generation=named_stats,
-                chart_generation=chart_stats,
-            ),
-        }
-    )
 
 
-@app.route("/api/query/refine-chart", methods=["POST"])
-@require_basic_auth
-async def api_query_refine_chart():
-    """Refine an existing chart spec based on user instruction."""
-    settings = _load_all_ai_settings(get_db())
-    if not _query_page_enabled(settings):
-        return jsonify({"ok": False, "error": "Query page is unavailable."}), 404
-
-    payload = await request.get_json() or {}
-    current_spec = payload.get("chart_spec", "")
-    columns = payload.get("columns", [])
-    rows = payload.get("rows", [])
-    user_instruction = payload.get("instruction", "").strip()
-    thinking_level = _normalize_thinking_level(str(payload.get("thinking_level") or "off"))
-
-    if not current_spec:
-        return jsonify({"ok": False, "error": "No chart spec provided."}), 400
-    if not user_instruction:
-        return jsonify({"ok": False, "error": "No instruction provided."}), 400
-
-    # Use current row data as sample if available, otherwise empty list
-    sample_rows = rows[:20] if rows else []
-
-    trace_id = str(uuid.uuid4())
-    turn_id = str(uuid.uuid4())
-    model = settings.get("ai.model", "").strip()
-
-    # Emit trace start event
-    _emit_ai_helper_log_event(
-        event_name="query.turn.start",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model="",
-        thinking_level="off",
-        body=f"Chart refinement requested: {user_instruction}",
-        attrs={
-            "gen_ai.operation.name": "refine_chart",
-            "sobs.gen_ai.instruction": user_instruction,
-        },
-    )
-
-    chart_spec, chart_error, chart_stats = await _vanna_refine_chart_spec(
-        current_spec, columns, sample_rows, user_instruction, settings, thinking_level=thinking_level
-    )
-
-    # Emit chart refinement event with LLM call details
-    _emit_ai_helper_log_event(
-        event_name="query.chart.refined",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model="",
-        thinking_level="off",
-        body=chart_spec if chart_spec else chart_error,
-        severity="ERROR" if chart_error else "INFO",
-        attrs={
-            "gen_ai.operation.name": "refine_chart",
-            "gen_ai.usage.input_tokens": chart_stats.get("prompt_tokens", 0),
-            "gen_ai.usage.output_tokens": chart_stats.get("completion_tokens", 0),
-            "gen_ai.response.latency_ms": chart_stats.get("elapsed_ms", 0),
-            "sobs.gen_ai.instruction": user_instruction,
-        },
-    )
-
-    # Emit turn complete event
-    _emit_ai_helper_log_event(
-        event_name="query.turn.complete",
-        chat_id=trace_id,
-        turn_id=turn_id,
-        page="/query",
-        model=model,
-        guard_model="",
-        thinking_level="off",
-        body="Chart refinement completed",
-        severity="ERROR" if chart_error else "INFO",
-        attrs={
-            "gen_ai.operation.name": "refine_chart",
-        },
-    )
-
-    if chart_error:
-        return (
-            jsonify(
-                {
-                    "ok": False,
-                    "error": chart_error,
-                    "trace_id": trace_id,
-                }
-            ),
-            500,
-        )
-
-    return jsonify(
-        {
-            "ok": True,
-            "trace_id": trace_id,
-            "chart_spec": chart_spec,
-        }
-    )
 
 
-@app.route("/api/query/schema", methods=["GET"])
-@require_basic_auth
-async def api_query_schema():
-    """Return the schema context string used for LLM prompts."""
-    settings = _load_all_ai_settings(get_db())
-    if not _query_page_enabled(settings):
-        return jsonify({"ok": False, "error": "Query page is unavailable."}), 404
-    db = get_db()
-    runner = ChdbSqlRunner(db)
-    schema = await asyncio.to_thread(runner.get_schema_context)
-    return jsonify({"ok": True, "schema": schema})
 
 
 # ---------------------------------------------------------------------------
@@ -24508,130 +22532,12 @@ async def api_query_schema():
 # ---------------------------------------------------------------------------
 
 
-@app.route("/table-explorer")
-@require_basic_auth
-async def view_table_explorer():
-    """Render the visual database table explorer page."""
-    if not _query_page_enabled():
-        return (
-            "Table Explorer is unavailable until AI and guard settings are configured.",
-            404,
-        )
-    return await render_template("table_explorer.html")
 
 
-@app.route("/api/table-explorer/tables", methods=["GET"])
-@require_basic_auth
-async def api_table_explorer_tables():
-    """Return metadata for all allowed tables (name, column count, columns).
-
-    Response shape::
-
-        {
-            "ok": true,
-            "tables": [
-                {
-                    "name": "otel_logs",
-                    "column_count": 12,
-                    "columns": [
-                        {
-                            "name": "Timestamp",
-                            "type": "DateTime64(9)",
-                            "is_nullable": false,
-                            "is_primary_key": false,
-                            "is_sorting_key": true,
-                            "is_partition_key": false,
-                            "default_kind": "",
-                            "comment": ""
-                        },
-                        ...
-                    ]
-                },
-                ...
-            ]
-        }
-    """
-    if not _query_page_enabled():
-        return jsonify({"ok": False, "error": "Table Explorer is unavailable."}), 404
-    db = get_db()
-    runner = ChdbSqlRunner(db)
-    try:
-        tables = await asyncio.to_thread(runner.get_allowed_tables_info)
-        return jsonify({"ok": True, "tables": tables})
-    except Exception as exc:
-        return jsonify({"ok": False, "error": str(exc)}), 500
 
 
-@app.route("/api/table-explorer/table/<name>", methods=["GET"])
-@require_basic_auth
-async def api_table_explorer_table(name: str):
-    """Return detailed info for a single allowed table: columns, DDL, and sample rows.
-
-    Response shape::
-
-        {
-            "ok": true,
-            "table": "otel_logs",
-            "columns": [...],
-            "ddl": "CREATE TABLE otel_logs ...",
-            "sample": {
-                "columns": ["Timestamp", "ServiceName", ...],
-                "rows": [["2024-01-01 00:00:00", "my-svc", ...], ...]
-            }
-        }
-    """
-    if not _query_page_enabled():
-        return jsonify({"ok": False, "error": "Table Explorer is unavailable."}), 404
-
-    # Validate table is in the allowlist
-    if name not in _QUERY_ALLOWED_TABLES:
-        return jsonify({"ok": False, "error": f"Table '{name}' is not accessible."}), 403
-
-    db = get_db()
-    runner = ChdbSqlRunner(db)
-    try:
-        detail = await asyncio.to_thread(runner.get_table_detail, name)
-        return jsonify(
-            {
-                "ok": True,
-                "table": name,
-                "columns": detail["columns"],
-                "ddl": detail["ddl"],
-                "sample": detail["sample"],
-            }
-        )
-    except Exception as exc:
-        return jsonify({"ok": False, "error": str(exc)}), 500
 
 
-@app.route("/api/chart-types", methods=["GET"])
-@require_basic_auth
-async def api_chart_types():
-    """Return the catalog of available ECharts chart types with configurations."""
-    try:
-        import json as json_module
-
-        chart_types_path = os.path.join(os.path.dirname(__file__), "static", "echarts-chart-types.json")
-        if not os.path.exists(chart_types_path):
-            return (
-                jsonify(
-                    {
-                        "ok": False,
-                        "error": "Chart types catalog not found. Run: node scripts/extract-echarts-types.js",
-                    }
-                ),
-                404,
-            )
-
-        with open(chart_types_path, "r") as f:
-            catalog = json_module.load(f)
-
-        return jsonify({"ok": True, "data": catalog})
-    except Exception as e:
-        return (
-            jsonify({"ok": False, "error": f"Failed to load chart types: {str(e)}"}),
-            500,
-        )
 
 
 # ---------------------------------------------------------------------------
@@ -27563,6 +25469,9 @@ from routes.rum import rum_bp as _rum_bp  # noqa: E402
 from routes.tags import tags_bp as _tags_bp  # noqa: E402
 from routes.traces import traces_bp as _traces_bp  # noqa: E402
 from routes.apps import apps_bp as _apps_bp  # noqa: E402
+from routes.dashboards import dashboards_bp as _dashboards_bp  # noqa: E402
+from routes.query import query_bp as _query_bp  # noqa: E402
+from routes.reports import reports_bp as _reports_bp  # noqa: E402
 from routes.settings import settings_bp as _settings_bp  # noqa: E402
 
 app.register_blueprint(_errors_bp)
@@ -27575,6 +25484,9 @@ app.register_blueprint(_rum_bp)
 app.register_blueprint(_tags_bp)
 app.register_blueprint(_traces_bp)
 app.register_blueprint(_apps_bp)
+app.register_blueprint(_dashboards_bp)
+app.register_blueprint(_query_bp)
+app.register_blueprint(_reports_bp)
 app.register_blueprint(_settings_bp)
