+++ templates/base.html
@@ -725,7 +725,7 @@
         </a>
       </li>
       <li class="nav-item">
-        <a href="{{ url_for('list_dashboards') }}" class="nav-link {% if request.endpoint in ('list_dashboards', 'view_custom_dashboard', 'new_dashboard_form') %}active{% endif %}" title="Dashboards">
+        <a href="{{ url_for('dashboards.list_dashboards') }}" class="nav-link {% if request.endpoint in ('list_dashboards', 'view_custom_dashboard', 'new_dashboard_form') %}active{% endif %}" title="Dashboards">
           <i class="bi bi-speedometer2"></i><span class="sb-label">Dashboards</span>
         </a>
       </li>
@@ -776,7 +776,7 @@
       </li>
       {% if query_enabled %}
       <li class="nav-item">
-        <a href="{{ url_for('view_query') }}" class="nav-link {% if request.endpoint in ('view_query',) %}active{% endif %}" title="Query">
+        <a href="{{ url_for('query.view_query') }}" class="nav-link {% if request.endpoint in ('view_query',) %}active{% endif %}" title="Query">
           <i class="bi bi-database-check"></i><span class="sb-label">Query</span>
         </a>
       </li>
@@ -789,7 +789,7 @@
       </li>
       {% endif %}
       <li class="nav-item">
-        <a href="{{ url_for('list_reports') }}" class="nav-link {% if request.endpoint == 'list_reports' %}active{% endif %}" title="Reports">
+        <a href="{{ url_for('reports.list_reports') }}" class="nav-link {% if request.endpoint == 'list_reports' %}active{% endif %}" title="Reports">
           <i class="bi bi-bookmark-star"></i><span class="sb-label">Reports</span>
         </a>
       </li>
@@ -1063,7 +1063,7 @@
     setBusy(true);
 
     try {
-      var response = await fetch({{ url_for('api_create_report') | tojson }}, {
+      var response = await fetch({{ url_for('reports.api_create_report') | tojson }}, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({
