+++ templates/chart_editor_help.html
@@ -17,7 +17,7 @@
             <p class="text-secondary mb-0">Step-by-step setup for template charts and custom ECharts mode.</p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('list_dashboards') }}"
+           href="{{ url_for('dashboards.list_dashboards') }}"
            title="Back to Dashboards">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Dashboards</span>
         </a>
