+++ templates/custom_dashboards.html
@@ -33,7 +33,7 @@
         {% endif %}
       </div>
       <div class="card-footer d-flex justify-content-between align-items-center">
-          <a href="{{ url_for('view_custom_dashboard', dashboard_id=dash.id) }}" class="btn btn-sm btn-outline-primary"
+          <a href="{{ url_for('dashboards.view_custom_dashboard', dashboard_id=dash.id) }}" class="btn btn-sm btn-outline-primary"
             data-ai-action-id="dashboards.open.detail"
             data-ai-action-type="navigate"
             data-ai-handler="navigate"
@@ -42,7 +42,7 @@
             data-ai-confirm="false">
           <i class="bi bi-eye me-1"></i>Open
         </a>
-          <form method="post" action="{{ url_for('delete_dashboard', dashboard_id=dash.id) }}"
+          <form method="post" action="{{ url_for('dashboards.delete_dashboard', dashboard_id=dash.id) }}"
             data-confirm-title="Delete Dashboard"
             data-confirm-message="Delete dashboard '{{ dash.name }}'?"
             data-confirm-ok-label="Delete"
@@ -84,7 +84,7 @@
 <div class="modal fade" id="newDashboardModal" tabindex="-1">
   <div class="modal-dialog">
     <div class="modal-content">
-      <form method="post" action="{{ url_for('create_dashboard') }}">
+      <form method="post" action="{{ url_for('dashboards.create_dashboard') }}">
         <div class="modal-header">
           <h5 class="modal-title">New Dashboard</h5>
           <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
