+++ templates/query.html
@@ -10,7 +10,7 @@
   </div>
 {% endset %}
 {% set query_header_actions %}
-  <a href="{{ url_for('view_table_explorer') }}" class="btn btn-sm btn-outline-secondary" title="Table Explorer">
+  <a href="{{ url_for('query.view_table_explorer') }}" class="btn btn-sm btn-outline-secondary" title="Table Explorer">
     <i class="bi bi-grid-3x3 me-1"></i><span class="query-btn-label">Table Explorer</span>
   </a>
   <a href="{{ url_for('query_help') }}" class="btn btn-sm btn-outline-secondary page-help-btn" title="Help" aria-label="Query Help"><i class="bi bi-question-circle me-1"></i><span class="query-btn-label">Help</span></a>
