+++ templates/reports_help.html
@@ -17,7 +17,7 @@
             <p class="text-secondary mb-0">Save, load, share, import, and export filter configurations as named reports.</p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('list_reports') }}"
+           href="{{ url_for('reports.list_reports') }}"
            title="Back to Reports">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Reports</span>
         </a>
