+++ templates/settings.html
@@ -88,7 +88,7 @@
                     </p>
                     <div class="d-flex align-items-center justify-content-between mt-2">
                         <span></span>
-                        <a href="{{ url_for('list_dashboards') }}"
+                        <a href="{{ url_for('dashboards.list_dashboards') }}"
                            class="btn btn-sm btn-outline-success"
                            data-ai-action-id="settings.nav.dashboards"
                            data-ai-action-type="navigate"
