+++ templates/table_explorer_help.html
@@ -17,7 +17,7 @@
             <p class="text-secondary mb-0">Browse allowed observability table schemas and sample data. Read-only view.</p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_table_explorer') }}"
+           href="{{ url_for('query.view_table_explorer') }}"
            title="Back to Table Explorer">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Table Explorer</span>
         </a>
@@ -139,7 +139,7 @@
         </div>
         <div class="card-body">
             <p class="small text-secondary mb-0">
-                The Table Explorer is linked from the <a href="{{ url_for('view_query') }}"><strong>Query</strong></a> page
+                The Table Explorer is linked from the <a href="{{ url_for('query.view_query') }}"><strong>Query</strong></a> page
                 via the <em>Explore Tables</em> button in the page header.
                 Use it to discover column names and data shapes before writing a natural-language query or custom SQL.
             </p>
