+++ templates/web_traffic.html
@@ -528,7 +528,7 @@ function escHtml(v){return String(v||'').replace(/&/g,'&amp;').replace(/</g,'&lt
   if (!form || !saveBtn || !menu) return;
 
   function loadSavedReports() {
-    fetch({{ url_for('api_list_reports') | tojson }} + '?page_type=web_traffic')
+    fetch({{ url_for('reports.api_list_reports') | tojson }} + '?page_type=web_traffic')
       .then(r => r.json())
       .then(reports => {
         if (!reports.length) {
@@ -540,7 +540,7 @@ function escHtml(v){return String(v||'').replace(/&/g,'&amp;').replace(/</g,'&lt
                      '&to_ts=' + encodeURIComponent(rep.filters?.to_ts || '');
           return `<li><a class="dropdown-item small" href="${href}" title="${escHtml(rep.description || '')}">${escHtml(rep.name || 'Unnamed')}</a></li>`;
         }).join('') + 
-        `<li><a class="dropdown-item small text-secondary" href="{{ url_for('list_reports') }}"><i class="bi bi-gear me-1"></i>Manage reports</a></li>`;
+        `<li><a class="dropdown-item small text-secondary" href="{{ url_for('reports.list_reports') }}"><i class="bi bi-gear me-1"></i>Manage reports</a></li>`;
       })
       .catch(() => {
         menu.innerHTML = '<li><span class="dropdown-item text-secondary small">Failed to load reports</span></li>';
